+++ packages/ui/src/utils/TypeUtil.ts
@@ -38,7 +38,7 @@ export type TextType =
 
 export type TextAlign = 'center' | 'left' | 'right'
 
-export type SizeType = 'inherit' | 'xl' | '2lg' | 'lg' | 'md' | 'mdl' | 'sm' | 'xs' | 'xxs'
+export type SizeType = 'inherit' | 'xl' | 'lg' | 'md' | 'mdl' | 'sm' | 'xs' | 'xxs' | 'xxl'
 
 export type SpacingType =
   | '0'
@@ -46,6 +46,7 @@ export type SpacingType =
   | '2xl'
   | '3xl'
   | '4xl'
+  | '5xl'
   | '3xs'
   | '4xs'
   | 'l'
@@ -133,6 +134,7 @@ export type IconType =
   | 'github'
   | 'google'
   | 'helpCircle'
+  | 'id'
   | 'infoCircle'
   | 'lightbulb'
   | 'mail'