+++ examples/html-wagmi/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @examples/html-wagmi
 
+## 4.2.3-alpha.0
+
+### Patch Changes
+
+- feat: add support for coinbase smart accounts
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.2.3-alpha.0
+
+## 4.2.2
+
+### Patch Changes
+
+- feat: social login refactor. wagmi sendCalls support. refactor theme variables
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.2.2
+
 ## 4.2.1
 
 ### Patch Changes