+++ packages/scaffold-utils/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @web3modal/scaffold-utils
 
+## 4.2.3-alpha.0
+
+### Patch Changes
+
+- feat: add support for coinbase smart accounts
+
+- Updated dependencies []:
+  - @web3modal/core@4.2.3-alpha.0
+  - @web3modal/polyfills@4.2.3-alpha.0
+
+## 4.2.2
+
+### Patch Changes
+
+- feat: social login refactor. wagmi sendCalls support. refactor theme variables
+
+- Updated dependencies []:
+  - @web3modal/core@4.2.2
+  - @web3modal/polyfills@4.2.2
+
 ## 4.2.1
 
 ### Patch Changes