+++ packages/wagmi/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @web3modal/wagmi
 
+## 4.2.3-alpha.0
+
+### Patch Changes
+
+- feat: add support for coinbase smart accounts
+
+- Updated dependencies []:
+  - @web3modal/scaffold-utils@4.2.3-alpha.0
+  - @web3modal/scaffold@4.2.3-alpha.0
+  - @web3modal/siwe@4.2.3-alpha.0
+  - @web3modal/polyfills@4.2.3-alpha.0
+  - @web3modal/scaffold-react@4.2.3-alpha.0
+  - @web3modal/scaffold-vue@4.2.3-alpha.0
+
+## 4.2.2
+
+### Patch Changes
+
+- feat: social login refactor. wagmi sendCalls support. refactor theme variables
+
+- Updated dependencies []:
+  - @web3modal/polyfills@4.2.2
+  - @web3modal/scaffold@4.2.2
+  - @web3modal/scaffold-react@4.2.2
+  - @web3modal/scaffold-utils@4.2.2
+  - @web3modal/scaffold-vue@4.2.2
+  - @web3modal/siwe@4.2.2
+
 ## 4.2.1
 
 ### Patch Changes